+++ SadConsole/SplashScreens/PCBoot.cs
@@ -27,12 +27,12 @@ namespace SadConsole.SplashScreens
             cursor = new Components.Cursor(Surface) { DisableWordBreak = true };
 
             _endAnimation = new Instructions.InstructionSet() { RemoveOnFinished = true }
-                .Instruct(new Instructions.FadeTextSurfaceTint(new ColorGradient(Settings.ClearColor.SetAlpha(0), Settings.ClearColor.SetAlpha(255)), System.TimeSpan.FromSeconds(1)))
+                .Instruct(new Instructions.FadeTextSurfaceTint(new Gradient(Settings.ClearColor.SetAlpha(0), Settings.ClearColor.SetAlpha(255)), System.TimeSpan.FromSeconds(1)))
                 .Wait(System.TimeSpan.FromMilliseconds(0.500))
                 .Code((s, d) => { IsVisible = false; return true; });
 
             _startAnimation = new Instructions.InstructionSet { RemoveOnFinished = true }
-                .Instruct(new Instructions.FadeTextSurfaceTint(new ColorGradient(Settings.ClearColor.SetAlpha(255), Settings.ClearColor.SetAlpha(0)), System.TimeSpan.FromSeconds(1)))
+                .Instruct(new Instructions.FadeTextSurfaceTint(new Gradient(Settings.ClearColor.SetAlpha(255), Settings.ClearColor.SetAlpha(0)), System.TimeSpan.FromSeconds(1)))
                 .Wait(System.TimeSpan.FromMilliseconds(0.500))
                 .Code(() =>
                 {